Complete PRA logic
Issue: ZENKO-4773
