+++ tests/ctst/common/common.ts
@@ -1,6 +1,6 @@
 import { ListObjectVersionsOutput } from '@aws-sdk/client-s3';
 import { Given, setDefaultTimeout, Then, When } from '@cucumber/cucumber';
-import { Constants, S3, Utils } from 'cli-testing';
+import { Constants, Identity, IdentityEnum, S3, Utils } from 'cli-testing';
 import Zenko from 'world/Zenko';
 import { safeJsonParse } from './utils';
 import assert from 'assert';
@@ -11,6 +11,7 @@ import {
     runActionAgainstBucket,
     getObjectNameWithBackendFlakiness,
     verifyObjectLocation,
+    restoreObject,
 } from 'steps/utils/utils';
 import { ActionPermissionsType } from 'steps/bucket-policies/utils';
 
@@ -130,6 +131,18 @@ Given('{int} objects {string} of size {int} bytes',
         await addMultipleObjects.call(this, numberObjects, objectName, sizeBytes);
     });
 
+Given('{int} objects {string} of size {int} bytes on {string} site',
+    async function (this: Zenko, numberObjects: number, objectName: string, sizeBytes: number, site: string) {
+        this.resetCommand();
+
+        if (site === 'DR') {
+            Identity.useIdentity(IdentityEnum.ACCOUNT, `${Zenko.sites['source'].accountName}-replicated`);
+        } else {
+            Identity.useIdentity(IdentityEnum.ACCOUNT, Zenko.sites['source'].accountName);
+        }
+        await addMultipleObjects.call(this, numberObjects, objectName, sizeBytes);
+    });
+
 Given('{int} objects {string} of size {int} bytes with user metadata {string}',
     async function (this: Zenko, numberObjects: number, objectName: string, sizeBytes: number, userMD: string) {
         await addMultipleObjects.call(this, numberObjects, objectName, sizeBytes, userMD);
@@ -220,17 +233,7 @@ Given('a transition workflow to {string} location', async function (this: Zenko,
 });
 
 When('i restore object {string} for {int} days', async function (this: Zenko, objectName: string, days: number) {
-    const objName = getObjectNameWithBackendFlakiness.call(this, objectName) ||  this.getSaved<string>('objectName');
-    this.resetCommand();
-    this.addCommandParameter({ bucket: this.getSaved<string>('bucketName') });
-    this.addCommandParameter({ key: objName });
-    const versionId = this.getSaved<Map<string, string>>('createdObjects')?.get(objName);
-    if (versionId) {
-        this.addCommandParameter({ versionId });
-    }
-    this.addCommandParameter({ restoreRequest: `Days=${days}` });
-    const result = await S3.restoreObject(this.getCommandParameters());
-    this.setResult(result);
+    await restoreObject.call(this, objectName, days);
 });
 
 // wait for object to transition to a location or get restored from it
