+++ tests/ctst/features/pra.feature
@@ -8,7 +8,7 @@ Feature: PRA operations
     Scenario Outline: PRA
     Given a "<versioningConfiguration>" bucket
     And a transition workflow to "e2e-cold" location
-    And <objectCount> objects "obj" of size <objectSize> bytes
+    And <objectCount> objects "obj" of size <objectSize> bytes on "Primary" site
     Then object "obj-1" should be "transitioned" and have the storage class "e2e-cold"
     And object "obj-2" should be "transitioned" and have the storage class "e2e-cold"
     And dmf volume should contain <objectCount> objects
@@ -16,13 +16,27 @@ Feature: PRA operations
     Then the DR source should be in phase "Running"
     And the DR sink should be in phase "Running"
     Given access keys for the replicated account
-    Then object "obj-1" should be "transitioned" and have the storage class "e2e-cold" on DR site
+
+    Then object "obj-1" should be "transitioned" and have the storage class "e2e-cold" on "DR" site
+    And object "obj-2" should be "transitioned" and have the storage class "e2e-cold" on "DR" site
+    Given <objectCount> objects "obj2" of size <objectSize> bytes on "Pimary" site
+    Then object "obj2-1" should be "transitioned" and have the storage class "e2e-cold" on "Primary" site
+    And object "obj2-2" should be "transitioned" and have the storage class "e2e-cold" on "Primary" site
+    Then object "obj2-1" should be "transitioned" and have the storage class "e2e-cold" on "DR" site
+    And object "obj2-2" should be "transitioned" and have the storage class "e2e-cold" on "DR" site
+    When i restore object "obj-1" for 2 days on "Primary" site
+    Then object "obj-1" should be "restored" and have the storage class "e2e-cold" on "Primary" site
+    And object "obj-1" should be "transitioned" and have the storage class "e2e-cold" on "DR" site
+    When i restore object "obj2-1" for 2 days on "DR" site
+    Then object "obj2-1" should be "restored" and have the storage class "e2e-cold" on "DR" site
+    And object "obj2-1" should be "transitioned" and have the storage class "e2e-cold" on "Primary" site
+
     And the kafka DR volume exists
     When I uninstall DR
     Then the DR custom resources should be deleted
 
     Examples:
     | versioningConfiguration | objectCount | objectSize |
-    |           Non versioned |           2 |        100 |
+    |               Versioned |           2 |        100 |
